django/requirements.txt
Django==5.0
djangorestframework==3.15.2
fastapi
uvicorn
requests
sentencepiece